Add self-joined users and guests to the candidate mentions in file rooms
File rooms were available only to users with direct access to the file,
so those were the users that could be mentioned. Now file rooms are
publicly available if the file is shared by link, so the mentions now
include all the users with direct access to the file like before plus
all the self-joined users and guests currently in the room.

Note, however, that self-joined users and guests can still not mention
users with direct access to the file that have not joined the room yet,
but only those that joined the room already (plus any participant that
is currently in the room).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>