Add basic acceptance tests for the Talk sidebar in the public share page
The "I see that the current page is the shared link I wrote down" step
is defined in the server but needs to be extended in Talk to also set
(in the tests) the Talk sidebar as the ancestor of the chat view.

In order to extend a step the whole Context that it belongs to must be
extended. Note, however, that the step in the child Context should not
include any "Given", "When" or "Then" annotation, as it would clash with
the step defined in the parent Context (the "Override" annotation is
simply informative, as it is ignored by Behat); as long as the method
has the same signature the step from the child Context will be used
instead if the step from the parent Context.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>