Do not add system message for self joined users to file rooms
When a user with access to a file joins its file room the user is added
as a persistent participant, so a "XXX joined the conversation" system
message is shown. However, if a user does not have direct access to a
file and joins its room the user is not added as a persistent
participant, so the system message should not be shown in that case.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>