Add required "patch: true" option by default
In order to set the room name by calling "save('name', name, options)"
the options need to include "patch: true", so instead of requiring the
callers to explicitly add it the option is now added by default.

This also adds some sanity checks to ensure that "save()" is always used
as expected.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>