Sync rooms instead of fetching the model after renaming a room
When a room is renamed the external signaling server sends a message
that causes all the rooms to be fetched again automatically, so there is
no need to explicitly fetch the room again. Therefore, now "syncRooms"
is used instead of an explicit fetch; "syncRooms()" immediately fetches
all the the rooms again when the internal signaling server is used, but
waits until the message to fetch all the rooms again is sent when the
external signaling server is used.

The drawback of this change is that now, when the internal signaling
server is used, all the rooms are fetched again instead of the one that
was changed, but that should not be a big issue given that renaming a
room is a rather infrequent action and that all the rooms are fetched
anyway every ten seconds.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>