Do not mark the password field as required
An empty password can be sent to remove the current password of the
room, which does not match with the "required" semantics. Moreover, the
CSS styles show a red border in invalid input fields, so the red border
was shown too when removing the password despite being a valid
operation.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>