Fix keyboard navigation in tab headers
The tab headers contain a link that was focusable using keyboard
navigation. However, the CSS styles that highlight the tab header were
associated to the whole tab header, so even if the link was focused the
tab header was not highlighted.

Now the link is no longer focusable and the tab header itself is the
element that gets the focus, which causes it to be highlighted when
navigating with the keyboard. Also, the same selection behaviour applied
when the tab header is clicked is now applied when pressing space or
enter.

Finally, as the tab header can not be selected again when it is already
selected the tab header is not focusable either when it is already
selected.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>