Fix select2 containers not highlighted when focused
The select2 containers have the same size as the "oca-spreedme-add-XXX"
wrappers that contain them, so the borders can be moved from those
wrappers to the containers themselves. This makes possible to set a
stronger border when the select2 container is active/focused.

Besides that now the text is stronger too when the select2 container is
focused to show that more clearly.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>