Fix edit button in EditableTextLabel not focusable
The edit button was not an actual button but a span, so it was ignored
in keyboard navigation. Moreover, the opacity was applied to its parent
element instead of to the button itself, so it was not highlighted when
focused. Also, after the input is hidden now the focus is given back to
the edit button.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>