Fix password button not focusable
The password button was not an actual button but a span, so it was
ignored in keyboard navigation. Also, after the password menu is hidden
now the focus is given back to the password button.

The button was also wrapped in a div with the "menutoggle" CSS class,
which causes the menu to be hidden when clicking again on the toggle.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>