Do not defer rendering CallInfoView again until the menu is inactive
In Nextcloud 17 "OC._currentMenu" was removed, so "renderWhenInactive()"
always renders immediately. Instead of fixing that the method was
completly removed, as keeping the menu open instead of rendering again
the view does not provide much value (if the room is no longer public
the password will be unusable anyway, and having the password menu open
when another moderator sets or removes the password should not happen
too often).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>