Remove no longer needed parent element of "Copy link" button
The "share-link-options" element acts as a spacer to move the room
moderation menu button to the right. However, given that the call button
width is limited to the 50% of its parent flex element and that element
has the full width of the sidebar the same effect can be achieved by
setting "margin-left: auto" on the room moderation button.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>