Generate "hashedSessionId" attribute in rooms from "sessionId"
In the chat guest users are identified by their "sessionId". The
"sessionId" is larger than the size of the "actorId" column in the
database, though, so the ID stored in the database and returned in chat
messages is a hashed version instead. For convenience, the room model
now provides the hashed version of the "sessionId" attribute, which can
be used to check if a hashed sessionId matches the current guest
participant.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>