Fix self avatar size for guests
The avatars were not usually 128px width and height because
"imageplaceholder" uses the height of the element and, if it is not
available (for example, because the element was never shown), it uses
the given size. In most cases the ChatView was rendered after being
shown, but in some special cases it could be rendered before being shown
and then shown, which caused the large avatars.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>