Set the automatic read marker before the waiting request, not afterwards
This false set the read marker on new messages although you
navigated away to a different chat already. So we removed this
and instead update the read marker before your next waiting.
So when you are still there, it will just have a wrong read
marker for the time until your next request starts, while it will
not update the value, when you actually left the chat already.

Signed-off-by: Joas Schilling <coding@schilljs.com>