Fix setting again the same lobby state in the UI
Sending a request to the server to set the same lobby state is valid,
but unneeded. Moreover, the CallInfoView is rendered again only when the
lobby state changes, so if the same state is set again a loading spinner
will be shown instead of the radio button until something else changes
in the room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>