Override the current user when getting the room for a public share page
The public share page uses the incognito mode, so "OC.getCurrentUser()"
never returns a user, even if the user is actually logged in. However,
Talk API requests honour the actual current user, so the Talk UI needs
to honour it as well.

Moreover, when the external signaling server is used, the Talk UI needs
to honour the current user not only to show the right user name and
avatar, but to be able to even connect with the external signaling
server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>