Fix guest avatars in public share page
The default avatar size in VideoView is 128px, but due to the limited
space available in the public share page it is reduced to 64px in the
CSS. However, as "imageplaceholder()" is called while the element is
still hidden it does not get the size set in CSS, but the one given to
the function. This causes a mismatch in the text size once the element
is shown with a expected size of 128px on an avatar of only 64px, so the
same size should be used in both cases.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>