Update the stored height of the new message field after cleaning it
When the new messsage field is cleaned after successfully sending a
message its height changes, so the stored height needs to be updated.
Otherwise when a keydown event is received again on the new message
field the old height would be used for the calculations, which would
cause the chat view to be wrongly scrolled after sending a message with
several lines.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>