Update stored height for new message field when chat view is resized
When the chat view is resized the new messsage field height may change,
so the stored height needs to be updated. Otherwise when a keydown event
is received again on the new message field the old height would be used
for the calculations, which would cause the chat view to be wrongly
scrolled if a message is currently in the new message field and its
height changed due to the resize.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>