Limit moderation menu width so it does not overflow the sidebar
The moderation menu width is based on its contents, and it is mainly
affected by the width of the datetime picker. The sidebar width depends
on the viewport width, and the default width of the datetime picker
causes the moderation menu to be wider than the sidebar and overflow it
when the window is below ~1300px.

Besides depending on the viewport width, the sidebar width is also
limited to a minimum and maximum size of 300px and 500px, so the
sidebar width is always 300px when the viewport width is < 1111px. Due
to this the minimum width of the datetime picker can be set to 155px to
ensure that the moderation menu will not overflow the sidebar when it is
300px, and use a dynamic width based on the viewport width when it is
between 1111px and 1315px, fixing its width at the default 210px once
the viepower is > 1315px.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>