Fix calendar icon in datetime picker partially hidding its input borders
When shown in a menu the datetime picker sizes and paddings are slightly
modified from its default style, which in some browsers can cause the
calendar icon to overlap the input and partially hide its borders due to
its background color. Inputs with icons are right padded in Talk, so the
input text will not appear below the calendar icon. Due to this the icon
background can be made transparent so it does not hide the borders
without having to fight with sizes and the slight rendering differences
of the layout depending on the browser.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>