Revert "Fix avatar container height during calls"
This reverts commit 6b04fb7ff9f8fbbc9cf87c647bda9eed9b10b52a.

Despite what was stated in "Fix avatar container height during calls"
the absolute position did not cause the avatar to have different height
when the contained div has a text or not; this was actually caused by
the "display: inline-block" in the avatar itself (although it was fixed
anyway by propagating the flex property).

However, the absolute position is actually needed to promote the avatar
(as this is done with a "top: 30%" rule) as well as to properly
distribute the non promoted users at the bottom of the video screen
(otherwise they use the full window width instead of only the video
screen width), but having both a flex display and an absolute position
breaks the layout (for example, when there are only two participants) in
a worse way than the original issue (having guest avatars at a slightly
higher position than the user avatar). Due to all this the original
"fix" is reverted.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>