Fix avatar container height during calls
The inline-block display causes the avatar container to have a different
height when the contained div has a text (guest avatars) or not (user
avatars, which use an image). The inline-block display is used with the
"text-align: center" in the parent container to center the avatar, so
now the "display: inline-block" property is removed and the avatar is
centered by means of its margins.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>