Fix guest avatars in Files app
In the public share page the avatars were always forced to 64px to
prevent some issues with "imageplaceholder()" (the avatar size is
hardcoded to the size of the element or, if the element is hidden, to
the size given to the function). In the Files app, however, the size of
the avatar changes dynamically when the user is promoted and unpromoted
between 128px and 64px. Due to this the hardcoded avatar size needs to
be overriden in the CSS rules to switch between 64px and 128px as
needed.

For consistency with the Files app, the promoted avatars in the public
share page are now 128px too.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>