Do not load Talk sidebar in public share page of folder shares
The Talk sidebar is only shown for file shares, so there is no need to
load it for folder shares. Moreover, this also prevents some of the
hacks used to show the Talk sidebar to mess with the layout used for
folders.

Note that the "Files_Sharing::loadAdditionalScripts" event only includes
the share since Nextcloud 17.0.1, so it is needed to handle the event
without parameters to keep compatibility with Nextcloud 17.0.0 (in that
case the Talk sidebar will not be loaded in any public share page until
the server is updated to 17.0.1 or later).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>