Fix scrolling position when new messages arrive while sidebar is hidden
The sidebar was hidden just by setting a width of 0, so the message list
was not really hidden, and if new messages were received they were added
to the list. As the message list width was 0 the visibility calculations
did not work as expected and the list was wrongly scrolled. Due to this
now the sidebar is actually hidden besides having a width of 0.

Note that, even with that change, the scroll position also needs to be
saved and restored when the sidebar is hidden and shown again, and the
message list needs to be reloaded when shown again to show the new
messages that may have been received while it was hidden.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>