Do not immediately join the conversation in Files app
Before as soon as the chat tab loaded if the file was shared the user
automatically joined the conversation. Now the conversation must be
explicitly joined by clicking a button in the chat tab. This prevents
accidentally joining a room or even creating them when opening the
details of a different file.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>