Use nextcloudci Docker image for node
The NPM major version was recently bumped in the node:lts Docker image,
so CI started to fail due to unclean JavaScript builds. The server seems
to require the previous version due to incompatible dependencies, so for
consistency the JavaScript files are still built and checked with the
previous version instead of updating them to the new one.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>