Fix DatetimePicker for guests
The DatetimePicker was loaded in the UI for guests, but as the needed
files were not included in "merged-guest.json" the UI for guests was
broken. As the DatetimePicker is never shown in the UI for guests now it
is loaded only in the main Talk UI used by logged in users.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>