Enable join button by default in public share pages
The join button was initially disabled in public share pages and enabled
once the Talk sidebar was fully shown. As the Talk sidebar is shown
using a transition showing the Talk sidebar had to be delayed after the
page load, as in some cases the transition was ignored and the sidebar
was immediately shown. When that happened, as there was no transition,
the join button was never enabled.

However, despite using a delay of one second in some cases (apparently
if the system load is high) the transition did not happen either.
Instead of increasing the delay even more, which does not guarantee
anything, now the join button is enabled since the beginning.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>