Fix moderation menu width when date picker is not shown
When the date picker is not shown the moderation menu width was based on
the checkboxes, which caused the menu item with the explanation about
the lobby to be quite narrow and stretch vertically to fit its contents.
Now the details width is based on the date picker width, even when it is
not shown, which provides more space for the text and prevents the menu
width from changing when the lobby is enabled or disabled.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>