Fix loading icon not shown on avatar
In some cases, for example during the initial connection, the loading
icon was not shown on the user avatar. The reason was that the
"avatar()" function adds a loading icon to the element while the image
is being fetched and removes it once done. Due to this, if the element
had a loading icon before it will be removed after the image is fetched,
and thus needs to be restored.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>