Send offer again until a connection is established
When the MCU is not used the peer with the larger ID sends an offer to
the peer with the smaller ID. However, sometimes the peer with the
larger ID may fail to notice a disconnection and not send an offer
again, so if the peer with the smaller ID does not get an offer in a
reasonable time it also sends an offer.

Before the offer was sent just once, so if the offer sent by the peer
with the smaller ID was not handled for some reason by the other peer
no reconnection happened. Now the offer is periodically resent until the
connection is established again.

Note that, besides improving reconnections, this also helps with the
initial connection, as if the initial offers are not handled now the
peer with the smaller ID will try to connect again and again (although
there are still "glare" issues if offers are crossed between peers).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>