Remove forced vertical scrollbar in shared screens area
"overflow-y: scroll" was added back in the day to restrict the scrollbar
only to the vertical position. However, that rule forces the scrollbar
to be always visible, even if it is not needed. Instead of changing it
to "overflow-x: hidden" the rule was just removed, as the size of the
shared screen is adjusted to the available space and thus no scrollbar
should be needed (and, if it is, it will be shown by the browser).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>