Fix shared screen not shown in Files app nor public share pages
The shared screens were hidden in the Files app and the public share
page, as sharing the screen is not possible from there (in the case of
the public share page due to a bug in the local video width, but that is
a different story :-P ). However, as other participants can open the
conversation in the main Talk UI and share their screen from there the
shared screens should be shown too in the Talk sidebar.

Note that in the public share authentication page the shared screens
were already shown, so now the common CSS rules are used in that case.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>