Hide button to share the screen in the public share page
Due to a bug in the local video width the button to share the screen is
not visible (except when other user has shared the screen). For
consistency with the Talk sidebar in the Files app (which does not show
the button in any case) and until the bug in the local video width is
fixed the share screen button is now hidden in the public share page.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>