Fix videos in Talk sidebar when a screen is shared
When a screen was shared and the video of another participant was
enabled the video overlapped the shared screens. Now the height of the
video containers is limited to the vertical space not used by the shared
screens to prevent the overlapping (like done in the main Talk UI).

The Talk sidebar in the public share authentication page uses the same
height as the main Talk UI, but in the Files app and the public share
page it is shorter, so specific rules need to be used in those cases.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>