Force regular rules on two participants with same level selectors
When there are just one or two participants in a call the
".participants-1" or ".participants-2" CSS classes are applied on the
main call element. Those classes are used to set specific rules for
videos, but only if screensharing is not active, in which case the
regular rules are used instead.

Although overriding the specific rules using the main call element works
now the specific rules are overriden using same level selectors
(".participants-X") instead. This is a preparatory step to apply all the
rules in "video.scss" only to children of the main call element.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>