Fix too broad CSS rules for video elements
Some CSS rules were directly applied to "video" elements, which caused
those rules to be applied to any "video" element in the Files app when
loaded by the Talk sidebar. To prevent the same issue to happen with
other rules using generic names (like ".videoContainer") now all the
rules in "video.scss" are applied in the Talk sidebar only to children
of "#call-container".

All the CSS rules for the Talk sidebar that override those defined in
"videos.scss" now include "#call-container" in their path to ensure that
they will be more specific than the ones defined in "videos.scss".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>