Add dummy CallView
For now it is simply a container for the "#videos" and "#screens"
elements that does nothing.

Note that this changes the markup in the main Talk UI (as those elements
are now wrapped by "#call-container"), although this makes it consistent
with the markup of the Talk sidebar without changing anything in
practice (as ".inCall", ".participants-1" and so on are still set in
"#app-content").

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>