Simplify getting the participant name
When the MCU is used "peer.nick" is always undefined (a "nickChanged"
event is periodically sent, but "peer.nick" is not set).

When the MCU is not used, once the remote offer or answer has been
handled (which happens just before "setRemoteDescription()" is called
and thus just before the "videoAdded" event is emitted), "peer.nick" is
the "nick" property passed to the SimpleWebRTC object created by the
remote participant. That property is set to the display name of the
current user, which for users is always a value (it fall backs to the
user ID if none is set) and for guests is always undefined.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>