Split setting the user ID and the participant name
The user ID was set when a new participant was found in the list of
participants and then cached to be used elsewhere. The only exception
was the handling of the "nick" event, in which the user ID is sent with
the event if the MCU is used (if the MCU is not used the event is only
triggered for guests so no user ID is set).

With this change the user ID is now set after the VideoView was shown.
Due to this "imageplaceholder()" now gets the height from the element
instead of using the given one, so it needs to be overriden in the CSS
rules to ensure that the right size is used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>