Fix using signaling before its settings are fetched
The signaling object is created after the signaling settings are
fetched. As the settings are fetched as soon as the application starts
it was assumed that the signaling object was always available when it
was going to be used, but that may not be the case if the request to
fetch the settings take longer than expected.

Now the signaling object is guarded behind an async function that
returns it if it is available or fetches the settings and then returns
the object once available. Also, if a new signaling object is tried to
be created while the settings are being fetched it will also wait until
the first request succeeded and return the same object in all cases.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>