Check whether the Talk sidebar is supported for a file or not
The Talk sidebar is supported for a file when it meets certain sharing
conditions. If the Talk sidebar is not supported then a message is shown
to the user in the tab to share the file first.

In some cases it is not possible to know if the Talk sidebar is
supported or not only from the FileInfo data and it is necessary to
query the server. In those cases a loading icon is shown until the
response from the server is received.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>