Force scroll bars in tab contents instead of in whole sidebar
Before, when the chat view was shown in the sidebar the sidebar height
grew to accomodate the full list of messages. Instead of that the chat
view should be limited to the available space in the tab content, and
the scroll bar should be provided by the list of messages itself.

In other tabs, like in the participants tab when there were a lot of
participants, something similar happened, and the conversation
information at the top of the sidebar as well as the tab headers were
scrolled when the participants were scrolled.

As the chat view tab has to provide its own scroll bars for consistency
now the scroll bars of all tabs are shown for their content instead of
for the whole sidebar. Note that if the tab provides some more specific
scroll bars (like done in the chat tab) those are the scroll bars shown
instead of the ones for the tab itself.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>