Show chat view in chat tab in Files app
When the Talk sidebar is supported for a file a message is shown to the
user to explicitly join the conversation, and once joined the chat view
is shown in the chat tab. If a different file is selected the previous
conversation is left (even if the conversation for the new file is not
joined).

The sidebar tabs are created again whenever the sidebar is opened, be it
when changing to a different file or even when opening the sidebar again
for the same file after closing it. As the file of the current
conversation needs to be known to leave the conversation or not when the
FileInfo changes (so it is left if it is opened in a different file, but
kept if opened in the same file) the file ID of the current conversation
needs to be stored outside of the view.

Finally, as the AdvancedInput in shown in the chat view but there is no
route in use "this.$route.name" needs to be guarded.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>