Force scroll bars in tab contents instead of in whole Files sidebar
Before, when the chat view was shown in the sidebar the sidebar height
grew to accomodate the full list of messages (or the chat view just
filled a section of the sidebar, depending on the browser). Instead of
that the chat view should grow to fill the available space in the tab
content, and the scroll bar should be provided by the list of messages
itself.

In order to do that it is necessary to modify the style of the tabs
content container to set "overflow: hidden". In Nextcloud 17 it was
possible to provide an additional set of CSS classes to be set in the
tabs content container when a tab was active with
"getTabsContainerExtraClasses", but unfortunately that is not currently
possible in Nextcloud 18. Therefore this needs to be done by explicitly
setting the desired CSS rules on the tabs content container when the
chat tab is the active tab (and removing it when it is no longer the
active tab).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>